% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candiSNP-package.R
\docType{package}
\name{candiSNP-package}
\alias{candiSNP}
\alias{candiSNP-package}
\title{candiSNP: Shiny app for finding regions enriched with candidate SNPs graphically}
\description{
Generates density plots from user-provided SNP data obtained from HTS. Candidate causative mutations, defined as SNPs causing non-synonymous changes in annotated coding regions are highlighted on the plots.
}
\author{
\strong{Maintainer}: Dan MacLean \email{dan.maclean@tsl.ac.uk}

}
\keyword{internal}
