% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{do_centromere_filter}
\alias{do_centromere_filter}
\title{removes points within +/- 0.5 Mb of the centromere in Arabidopsis plots in the app}
\usage{
do_centromere_filter(df, centromere_state, genome)
}
\arguments{
\item{df}{parsed SNPEff data}

\item{centromere_state}{"Hidden" or "Visible" as to whether centromere SNPs should be shown}

\item{genome}{SNPEff genome in use.}
}
\value{
dataframe
}
\description{
removes points within +/- 0.5 Mb of the centromere in Arabidopsis plots in the app
}
