% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{download_annotation}
\alias{download_annotation}
\title{download annoations from SNPEff}
\usage{
download_annotation(genomes = NULL)
}
\arguments{
\item{genomes}{character vector of annotation keys (genome names) as described in snpeff.config}
}
\value{
no value
}
\description{
download annoations from SNPEff
}
