% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_installed_genomes}
\alias{get_installed_genomes}
\title{list installed genomes}
\usage{
get_installed_genomes()
}
\value{
character vector of genomes available
}
\description{
list installed genomes
}
\examples{
 get_installed_genomes()
}
