% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_snp_type}
\alias{get_snp_type}
\title{determines the candiSNP snp type for each SNP from SNPEff data
adds a new column 'type'}
\usage{
get_snp_type(df)
}
\arguments{
\item{df}{dataframe of parsed SNPEff data}
}
\value{
dataframe
}
\description{
determines the candiSNP snp type for each SNP from SNPEff data
adds a new column 'type'
}
