% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{install_default_genomes}
\alias{install_default_genomes}
\title{install the default genome annotations from SNPEff}
\usage{
install_default_genomes()
}
\value{
No value
}
\description{
installs the following genome annotations from SNPEff:
     "athalianaTair9",
     "athalianaTair10",
     "gmax1.09v8",
     "grape",
     "maizeZmB73",
     "rice7"
}
\details{
these being the ones originally available in candiSNP
}
\examples{
 install_default_genomes()
}
